doc[zfoo]: 修改文档注释

+++ util/src/main/java/com/zfoo/util/captcha/gif/Quant.java
@@ -14,7 +14,7 @@
 package com.zfoo.util.captcha.gif;
 
 /**
- * @author jaysunxiao
+ * @author godotg
  * @version 3.0
  */
 public class Quant {
@@ -51,9 +51,9 @@ public class Quant {
 
     /* defs for freq and bias */
     protected static final int intbiasshift = 16; /* bias for fractions */
-    protected static final int intbias = (((int) 1) << intbiasshift);
+    protected static final int intbias = (1 << intbiasshift);
     protected static final int gammashift = 10; /* gamma = 1024 */
-    protected static final int gamma = (((int) 1) << gammashift);
+    protected static final int gamma = (1 << gammashift);
     protected static final int betashift = 10;
     protected static final int beta = (intbias >> betashift); /* beta = 1/1024 */
     protected static final int betagamma =
@@ -62,18 +62,18 @@ public class Quant {
     /* defs for decreasing radius factor */
     protected static final int initrad = (netsize >> 3); /* for 256 cols, radius starts */
     protected static final int radiusbiasshift = 6; /* at 32.0 biased by 6 bits */
-    protected static final int radiusbias = (((int) 1) << radiusbiasshift);
+    protected static final int radiusbias = (1 << radiusbiasshift);
     protected static final int initradius = (initrad * radiusbias); /* and decreases by a */
     protected static final int radiusdec = 30; /* factor of 1/30 each cycle */
 
     /* defs for decreasing alpha factor */
     protected static final int alphabiasshift = 10; /* alpha starts at 1.0 */
-    protected static final int initalpha = (((int) 1) << alphabiasshift);
+    protected static final int initalpha = (1 << alphabiasshift);
     /* radbias and alpharadbias used for radpower calculation */
     protected static final int radbiasshift = 8;
-    protected static final int radbias = (((int) 1) << radbiasshift);
+    protected static final int radbias = (1 << radbiasshift);
     protected static final int alpharadbshift = (alphabiasshift + radbiasshift);
-    protected static final int alpharadbias = (((int) 1) << alpharadbshift);
+    protected static final int alpharadbias = (1 << alpharadbshift);
     protected int alphadec; /* biased by 10 bits */
 
     /* Types and Global Variables
@@ -433,7 +433,7 @@ public class Quant {
         int bestpos, bestbiaspos, bestd, bestbiasd;
         int[] n;
 
-        bestd = ~(((int) 1) << 31);
+        bestd = ~(1 << 31);
         bestbiasd = bestd;
         bestpos = -1;
         bestbiaspos = bestpos;
